Refactoring and seperated App Class
