+++ example.py
@@ -1,29 +1,18 @@
-import sys
 from random import randint
-
-import pyglet
-from pyglet.gl import *
 from glm import vec3
 
+from ThreeDEngine.app import App
 from ThreeDEngine.options import Options
 from ThreeDEngine.camera import Camera
 from ThreeDEngine import draw
 
 
-class App(pyglet.window.Window):
+class Example(App):
     def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
         self.color = [randint(50, 255) for i in range(3)]
-
-        pyglet.clock.schedule_interval(self.schedule_callback, 1/Options.fps)
-        pyglet.app.run()
-        glClearColor(*Options.background_color)
-
-    def schedule_callback(self, dt):
-        self.update(dt)
-        self.clear()
-        glLineWidth(Options.stroke_width)
-        self.render()
+        # Declare your variables before super().__init__() is called
+        
+        super().__init__(*args, **kwargs)
 
     def update(self, dt):
         Camera.pos.x += 100 * dt
@@ -36,4 +25,4 @@ class App(pyglet.window.Window):
 
 
 if __name__ == "__main__":
-    window = App(*Options.window_size.to_tuple(), "3D iz kewl!")
+    window = Example(*Options.window_size.to_tuple(), "3D iz kewl!")
